# api/openapi.yaml
components:
  schemas:
...
    AccountLimits:
      additionalProperties: false
      properties:
        history_from:
          description: "The trips and charges that ended before it are neither listed nor counted in the statistics: they are kept, and shown again once the limit is lifted. null: none is hidden."
          format: date-time
          pattern: Z$
          type:
            - string
            - "null"
        unavailable:
          description: "The functions the account does not have. costs: the charges' costs, null in the charges and the statistics, and their entry refused (feature_unavailable). stats: the statistics split into intervals, refused (feature_unavailable); a period's totals stay. csv: the trips and charges as CSV files, refused (feature_unavailable); the account's full export stays. mqtt: an MQTT broker, whose setting is refused (feature_unavailable); one set before stays, and nothing is published to it."
          items:
            enum:
              - costs
              - stats
              - csv
              - mqtt
            type: string
          type: array
        unread_vehicles:
          description: "The account's vehicles the collector does not read: what was read of them stays, and is shown, but nothing new is, until the limit is lifted."
          items:
            format: uuid
            type: string
          type: array
      required:
        - history_from
        - unavailable
        - unread_vehicles
      type: object
...
    Credentials:
      additionalProperties: false
      properties:
        password:
          type: string
        username:
          type: string
      required:
        - username
        - password
      type: object
...
    Error:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - error
      type: object
    ErrorDetail:
      additionalProperties: false
      properties:
        code:
          description: Stable and machine-readable.
          enum:
            - invalid_parameter
            - invalid_body
            - unauthorized
            - invalid_credentials
            - cross_origin
            - not_found
            - method_not_allowed
            - unsupported_media_type
            - too_many_attempts
            - currency_in_use
            - without_position_taken
            - too_many_places
            - charge_has_cost
            - api_key_refused
            - instance_key
            - too_many_vehicles
            - feature_unavailable
            - broker_refused
            - insufficient_scope
            - rate_limited
            - internal
          type: string
        message:
          description: For humans; it may change.
          minLength: 1
          type: string
      required:
        - code
        - message
      type: object
...
    Session:
      additionalProperties: false
      properties:
        expires_at:
          description: The absolute end of the session. It ends sooner after 7 days without a request, or when the password changes.
          format: date-time
          pattern: Z$
          type: string
        limits:
          description: "What the account's offer hides, as of now; null: nothing, the whole history is shown. Only an instance that limits its accounts sets it."
          oneOf:
            - $ref: "#/components/schemas/AccountLimits"
            - type: "null"
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
        - expires_at
        - limits
      type: object
...
    User:
      additionalProperties: false
      properties:
        id:
          type: string
        username:
          description: Lowercase.
          type: string
      required:
        - id
        - username
      type: object
...
  securitySchemes:
    accessToken:
      description: "A personal access token, `rst_` then 43 characters, issued by `POST /tokens`: it reads only."
      scheme: bearer
      type: http
    session:
      description: The session cookie set by `POST /session`, when the instance is served over plain http.
      in: cookie
      name: runsten_session
      type: apiKey
    sessionSecure:
      description: "The same cookie when the instance is served over https (its redirect URI is https): `Secure`, and the `__Host-` prefix binds it to the host."
      in: cookie
      name: __Host-runsten_session
      type: apiKey
info:
  contact:
    name: Runsten
    url: https://github.com/runsten-app/runsten
  description: |-
    The session, the account's vehicles, their current state (Verdandi), their
    trips and charges (Urd) with the cost of each charge, their statistics, the account's
    settings (its currency, and its places with their tariffs), and the costs entered for
    charges.

    Conventions:

    - Units are in the field names: `_km`, `_m` (metres), `_kwh`, `_pct` (state of charge, in %), `_w`, `_kw`, `_kwh_per_100km`, `_s` (seconds), `_minor` (minor units of the account's currency, as cents: an integer), `price_per_kwh` (major units of the account's currency per kWh).
      Positions are WGS 84, in degrees.
    - Times are RFC 3339, in UTC.
    - An unknown value is `null`, never zero. Every key is always present.
    - Lists are `{"items": [...]}`, never `null`. Trips and charges are paged with an
      opaque cursor, newest first.
    - Errors are `{"error": {"code", "message"}}`: the code is stable, the message is
      for humans and may change. A path under `/api/` that serves no route answers 404
      `not_found`, and a method that a path does not accept answers 405
      `method_not_allowed` with an `Allow` header.
    - Every response carries `Cache-Control: no-store`, `X-Content-Type-Options: nosniff`
      and `Referrer-Policy: no-referrer`.

    Every route but `POST /session` requires the session cookie it sets, or, to read, a
...
  license:
...
    name: AGPL-3.0-or-later
...
openapi: 3.1.0
paths:
...
  /session:
    delete:
      description: Deletes the session and clears its cookie.
      operationId: deleteSession
      responses:
        "204":
          description: Signed out.
          headers:
            Set-Cookie:
              schema:
                description: Clears the session cookie (Max-Age=0).
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`cross_origin`: a request that changes state, from another origin. `insufficient_scope`: this operation requires a session, never an access token."
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
      summary: Sign out
      tags:
        - session
    get:
      operationId: getSession
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
          description: The session of the cookie.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`insufficient_scope`: this operation requires a session, never an access token."
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
      summary: The current session
      tags:
        - session
    post:
      description: Checks the credentials and sets the session cookie. After 10 failures in 15 minutes for a username or a client address, attempts are refused until the end of the window.
      operationId: createSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
          description: Signed in.
          headers:
            Set-Cookie:
              schema:
                description: "The session cookie, runsten_session (or __Host-runsten_session over https): HttpOnly, SameSite=Lax, Path=/, until the session's absolute expiry."
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`invalid_body`: not a single JSON object with username and password, an unknown field, or more than 4 KiB."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`invalid_credentials`: unknown username or wrong password (the same answer for both)."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`cross_origin`: a request that changes state, from another origin."
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unsupported_media_type`: the body is not `application/json`."
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`too_many_attempts`: too many failed logins."
          headers:
            Retry-After:
              description: Seconds to wait before the next attempt.
              required: true
              schema:
                minimum: 1
                type: integer
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - {}
      summary: Sign in
      tags:
        - session
...
servers:
...